==> Step2_Assign_Tone.ipynb <==
# Task 2: Assign Respond Tone for Target Character's Lines

from transformers import AutoTokenizer, AutoModel
import torch
import pandas as pd
from tqdm import tqdm

# read in csv
df_chigiri = pd.read_csv('data/chigiri_pairs_new.csv')

# load BERT tokeizeer and model
tokenizer = AutoTokenizer.from_pretrained("cl-tohoku/bert-base-japanese", use_fast=False)
bert_model = AutoModel.from_pretrained("cl-tohoku/bert-base-japanese")
bert_model.eval()

def get_bert_embedding(text):
    """
    Get the bert embedding of the given text. Use mean pooling method to 
    keep the embedding vector the same size.
    
    Args:
        text: Japanese text to be embedded
        
    Return:
        embedded_vector: 768-d embedded_vector of a sentence
    """
    with torch.no_grad():
        # tokenize and get embedding vector for each token
        inputs = tokenizer(text, return_tensors="pt", truncation=True, max_length=128)
        outputs = bert_model(**inputs)
        # [batch_size, seq_len, hidden_size]
        hidden_states = outputs.last_hidden_state  
        
        # identify padding and make it into 0
        mask = inputs["attention_mask"].unsqueeze(-1).expand(hidden_states.size())
        masked_embeddings = hidden_states * mask
        
        # do mean pooling to get an average of the tokens
        # having 768-d embedding vectors for all sentences
        summed = torch.sum(masked_embeddings, dim=1)
        counts = torch.clamp(mask.sum(dim=1), min=1e-9)
        mean_pooled = summed / counts
    return mean_pooled.squeeze().numpy()

def sentence_embedding(df, text_column="chigiri_line", output_column="embedding"):
    """
    Embedding function for a whole dataframe. Add a new column for embedding vectors.
    
    Args:
        df: dataframe with text 
        text_column: column name of where text existed
        output_column: column name of where the embedded vectors should be
        
    Return:
        df: updated dataframe with embedded vecotr.
    """
    embeddings = []
    for text in tqdm(df[text_column]):
        if pd.isna(text) or text.strip() == "":
            embeddings.append(None)
        else:
            emb = get_bert_embedding(text)
            embeddings.append(emb)
    df[output_column] = embeddings
    return df

# do embedding for chigiri's dataset
df_chigiri = sentence_embedding(df_chigiri, text_column="chigiri_line")

# save the result for backup
# df_chigiri.to_pickle("chigiri_with_embedding.pkl")

from sklearn.cluster import KMeans
import numpy as np

# get out all embedding vectors
valid_embeddings = df_chigiri["embedding"].dropna().tolist()
X = np.stack(valid_embeddings)

# seperate them into 10 clusters
kmeans = KMeans(n_clusters=10, random_state=42)
df_chigiri.loc[df_chigiri["embedding"].notna(), "tone_cluster"] = kmeans.fit_predict(X)


# print out the first 10 sentences to check
for i in range(10): 
    print(f"\n🟦 Cluster {i} 語氣句子:")
    display(df_chigiri[df_chigiri["tone_cluster"] == i]["chigiri_line"].head(10))

# assign name for each cluster
dict_group_tone = {
    0: "テンション高めにふざけて・ネタっぽく", 
    1: "怒っている・キレている・強気で挑発的に",
    2: "言いたくない・感情を隠す・葛藤をにじませる", 
    3: "普通の会話、軽く返す・冷静にリアクション", 
    4: "皮肉っぽく・ぶっきらぼう・子供っぽく拗ねて", 
    5: "悔しさをにじませつつ本音で熱く", 
    6: "焦り・動揺・緊張した様子で", 
    7: "短く叫ぶ・反応を返す・テンポ速く強めに",
    8: "冷静に説明する・論理的に・分析的に",
    9: "感情爆発・強く叫ぶ・悔しさや怒りをぶつける", 
}

def assign_tone(df, cluster_to_tone, cluster_column="tone_cluster", output_column="chigiri_tone"):
    """
    Update the assigned tone to the original dataframe.
    
    Args:
        df: dataframe with text after clustering
        cluster_to_tone (dict): assigned name for each cluster
        cluster_column: name of the column where the cluster information existed
        output_column: name fo the new column
    
    """
    df[output_column] = df[cluster_column].map(cluster_to_tone)
    return df

df_chigiri_tone = assign_tone(df_chigiri, dict_group_tone)

df_chigiri_tone.head()

df_chigiri_tone = df_chigiri_tone.drop(columns=["embedding"])
df_chigiri_tone.to_csv("data/chigiri_train_w_tone_.csv")

Then, I manually check the assigned tone to make sure there isn't huge mistake.
